% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as.LAS}
\alias{as.LAS}
\title{data.frame of points in this package to LAS object of lidR package conversion.}
\usage{
as.LAS(df)
}
\description{
Convert data.frame of points (received from PointDB or PointCloud) to \link[lidR]{LAS} in lidR package.
}
\seealso{
\link[lidR]{LAS}
}
\author{
woellauer
}
