% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as.speclib}
\alias{as.speclib}
\title{RasterStack to Speclib conversion}
\usage{
as.speclib(rasterStack, na = NULL)
}
\description{
Convert RasterStack (received from RasterDB) to \link{speclib} for usage of functionality in \link{hsdar-package}.
}
\details{
For each contained RasterLayer attributes wavelength and fwhm need to be present (e.g. RasterStack created by \link{RasterDB}).

Speclib can be converted back to Raster(Brick) by \link{HyperSpecRaster} ( e.g. r <- HyperSpecRaster(sp) ).
}
\seealso{
\code{\link{speclib}} \link{hsdar-package}
}
\author{
woellauer
}
