% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointDB_functions.R
\name{contains_data}
\alias{contains_data}
\title{Tests if a data.frame of point-data contains at least one point.}
\usage{
contains_data(df)
}
\arguments{
\item{df}{data.frame of points}
}
\value{
TRUE or FALSE
}
\description{
Tests if a data.frame of point-data contains at least one point.
}
\author{
woellauer
}
