% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{convert_Polygon_to_matrix}
\alias{convert_Polygon_to_matrix}
\title{Convert polygon of class Polygon to coordinate matrix.}
\usage{
convert_Polygon_to_matrix(plgn)
}
\description{
(mostly for internal use) Preperation of polygon data for PointDB processing function.
}
\seealso{
\link{PointDB} \link{convert_Polygons_to_matrix} \link{convert_SpatialPolygons_to_named_matrix_list} \link{convert_SpatialPolygonsDataFrame_to_named_matrix_list}
}
\author{
woellauer
}
