% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{convert_SpatialPolygonsDataFrame_to_named_matrix_list}
\alias{convert_SpatialPolygonsDataFrame_to_named_matrix_list}
\title{Convert polygons of SpatialPolygonsDataFrame to named list of coordinate matrices.}
\usage{
convert_SpatialPolygonsDataFrame_to_named_matrix_list(areas)
}
\description{
Only first ring of each polygon is converted.
}
\details{
(mostly for internal use) Preperation of polygon data for PointDB processing function.
}
\seealso{
\link{PointDB} \link{convert_Polygon_to_matrix} \link{convert_Polygons_to_matrix} \link{convert_SpatialPolygons_to_named_matrix_list}
}
\author{
woellauer
}
