% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDAT.R
\name{read.rdat}
\alias{read.rdat}
\title{Read rDat-file}
\usage{
read.rdat(filename)
}
\description{
Reads rPointDB-package transfer format into R object.
}
\details{
If file contains raster data it is returned as RasterLayer or RasterStack.
}
\author{
woellauer
}
