% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointDB_functions.R
\name{statistics_data}
\alias{statistics_data}
\title{Calculates some statistical data of a set of points.}
\usage{
statistics_data(df)
}
\arguments{
\item{df}{data.frame of points}
}
\value{
named list statistical values
}
\description{
Calculates some statistical data of a set of points.
}
\author{
woellauer
}
